requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn
diffusers
transformers
accelerate
safetensors
scipy
gradio
torch
